Split the Dataset implementation using mixins
